Move enumerators to EventStore.Core.Services.Transport.Enumerators namespace
