+++ src/EventStore.Core.Tests/Services/Transport/Grpc/StreamRevisionTests.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using EventStore.Core.Services.Transport.Grpc;
+using EventStore.Core.Services.Transport.Common;
 using NUnit.Framework;
 
 namespace EventStore.Core.Tests.Services.Transport.Grpc {
