+++ src/EventStore.Core/Services/Transport/Common/HashCode.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace EventStore.Core.Services.Transport.Grpc {
+namespace EventStore.Core.Services.Transport.Common {
 	public struct HashCode {
 		private readonly int _value;
