+++ src/EventStore.Core/Services/Transport/Enumerators/Enumerator.AllSubscription.cs
@@ -9,9 +9,10 @@ using EventStore.Core.Data;
 using EventStore.Core.Messages;
 using EventStore.Core.Messaging;
 using EventStore.Core.Services.Storage.ReaderIndex;
+using EventStore.Core.Services.Transport.Common;
 using Serilog;
 
-namespace EventStore.Core.Services.Transport.Grpc {
+namespace EventStore.Core.Services.Transport.Enumerators {
 	partial class Enumerator {
 		public class AllSubscription : IAsyncEnumerator<ReadResponse> {
 			private static readonly ILogger Log = Serilog.Log.ForContext<AllSubscription>();
