+++ src/EventStore.Core/Services/Transport/Enumerators/Enumerator.ReadAllBackwards.cs
@@ -8,8 +8,9 @@ using EventStore.Core.Bus;
 using EventStore.Core.Data;
 using EventStore.Core.Messages;
 using EventStore.Core.Messaging;
+using EventStore.Core.Services.Transport.Common;
 
-namespace EventStore.Core.Services.Transport.Grpc {
+namespace EventStore.Core.Services.Transport.Enumerators {
 	partial class Enumerator {
 		public class ReadAllBackwards : IAsyncEnumerator<ReadResponse> {
