+++ src/EventStore.Core/Services/Transport/Enumerators/ReadResponse.cs
@@ -1,6 +1,7 @@
 ﻿using EventStore.Core.Data;
+using EventStore.Core.Services.Transport.Common;
 
-namespace EventStore.Core.Services.Transport.Grpc;
+namespace EventStore.Core.Services.Transport.Enumerators;
 
 public abstract class ReadResponse {
 	public class EventReceived: ReadResponse {
