+++ src/EventStore.Core/Services/Transport/Grpc/AnyStreamRevision.cs
@@ -1,4 +1,5 @@
 using System;
+using HashCode = EventStore.Core.Services.Transport.Common.HashCode;
 
 namespace EventStore.Core.Services.Transport.Grpc {
 	public struct AnyStreamRevision : IEquatable<AnyStreamRevision> {
