+++ src/EventStore.Core/Services/Transport/Grpc/PersistentSubscriptions.Read.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -13,6 +12,7 @@ using EventStore.Core.Bus;
 using EventStore.Core.Data;
 using EventStore.Core.Messages;
 using EventStore.Core.Messaging;
+using EventStore.Core.Services.Transport.Common;
 using EventStore.Plugins.Authorization;
 using Google.Protobuf;
 using Grpc.Core;
