+++ src/EventStore.Core/Services/Transport/Grpc/ReadReq.cs
@@ -1,5 +1,5 @@
 using System;
-using EventStore.Core.Services.Transport.Grpc;
+using EventStore.Core.Services.Transport.Common;
 
 // ReSharper disable CheckNamespace
 namespace EventStore.Client.Streams {
@@ -25,18 +25,18 @@ namespace EventStore.Client.Streams {
 					}
 
 					partial class AllOptions {
-						internal Core.Services.Transport.Grpc.Position ToPosition() => AllOptionCase switch {
-							AllOptionOneofCase.End => Core.Services.Transport.Grpc.Position.End,
-							AllOptionOneofCase.Start => Core.Services.Transport.Grpc.Position.Start,
-							AllOptionOneofCase.Position => new Core.Services.Transport.Grpc.Position(Position.CommitPosition,
+						internal Core.Services.Transport.Common.Position ToPosition() => AllOptionCase switch {
+							AllOptionOneofCase.End => Core.Services.Transport.Common.Position.End,
+							AllOptionOneofCase.Start => Core.Services.Transport.Common.Position.Start,
+							AllOptionOneofCase.Position => new Core.Services.Transport.Common.Position(Position.CommitPosition,
 								Position.PreparePosition),
 							_ => throw new InvalidOperationException()
 						};
 
-						internal Core.Services.Transport.Grpc.Position? ToSubscriptionPosition() => AllOptionCase switch {
-							AllOptionOneofCase.End => Core.Services.Transport.Grpc.Position.End,
+						internal Core.Services.Transport.Common.Position? ToSubscriptionPosition() => AllOptionCase switch {
+							AllOptionOneofCase.End => Core.Services.Transport.Common.Position.End,
 							AllOptionOneofCase.Start => null,
-							AllOptionOneofCase.Position => new Core.Services.Transport.Grpc.Position(Position.CommitPosition,
+							AllOptionOneofCase.Position => new Core.Services.Transport.Common.Position(Position.CommitPosition,
 								Position.PreparePosition),
 							_ => throw new InvalidOperationException()
 						};
