+++ src/EventStore.Core/Services/Transport/Grpc/Streams.Read.cs
@@ -4,9 +4,14 @@ using System.Linq;
 using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
+using EventStore.Client;
 using EventStore.Client.Streams;
+using EventStore.Core.Data;
 using EventStore.Core.Metrics;
 using EventStore.Core.Services.Storage.ReaderIndex;
+using EventStore.Core.Services.Transport.Common;
+using EventStore.Core.Services.Transport.Enumerators;
+using Google.Protobuf;
 using Grpc.Core;
 using CountOptionOneofCase = EventStore.Client.Streams.ReadReq.Types.Options.CountOptionOneofCase;
 using FilterOptionOneofCase = EventStore.Client.Streams.ReadReq.Types.Options.FilterOptionOneofCase;
@@ -247,7 +252,7 @@ namespace EventStore.Core.Services.Transport.Grpc {
 		private static ReadResp ConvertReadResponse(ReadResponse readResponse, ReadReq.Types.Options.Types.UUIDOption uuidOption) {
 			return readResponse switch {
 				ReadResponse.EventReceived eventReceived => new ReadResp {
-					Event = Enumerator.ConvertToReadEvent(uuidOption, eventReceived.Event)
+					Event = ConvertToReadEvent(uuidOption, eventReceived.Event)
 				},
 				ReadResponse.SubscriptionConfirmed subscriptionConfirmed => new ReadResp {
 					Confirmation = new ReadResp.Types.SubscriptionConfirmation {
@@ -301,6 +306,54 @@ namespace EventStore.Core.Services.Transport.Grpc {
 				default:
 					throw new ArgumentException($"Unknown read response exception type: {readResponseEx.GetType().Name}", nameof(readResponseEx));
 			}
+		}
+
+		private static ReadResp.Types.ReadEvent.Types.RecordedEvent ConvertToRecordedEvent(
+			ReadReq.Types.Options.Types.UUIDOption uuidOption, EventRecord e, long? commitPosition,
+			long? preparePosition) {
+			if (e == null) return null;
+			var position = Position.FromInt64(commitPosition ?? -1, preparePosition ?? -1);
+			return new ReadResp.Types.ReadEvent.Types.RecordedEvent {
+				Id = uuidOption.ContentCase switch {
+					ReadReq.Types.Options.Types.UUIDOption.ContentOneofCase.String => new UUID {
+						String = e.EventId.ToString()
+					},
+					_ => Uuid.FromGuid(e.EventId).ToDto()
+				},
+				StreamIdentifier = e.EventStreamId,
+				StreamRevision = StreamRevision.FromInt64(e.EventNumber),
+				CommitPosition = position.CommitPosition,
+				PreparePosition = position.PreparePosition,
+				Metadata = {
+					[Constants.Metadata.Type] = e.EventType,
+					[Constants.Metadata.Created] = e.TimeStamp.ToTicksSinceEpoch().ToString(),
+					[Constants.Metadata.ContentType] = e.IsJson
+						? Constants.Metadata.ContentTypes.ApplicationJson
+						: Constants.Metadata.ContentTypes.ApplicationOctetStream
+				},
+				Data = ByteString.CopyFrom(e.Data.Span),
+				CustomMetadata = ByteString.CopyFrom(e.Metadata.Span)
+			};
+		}
+
+		private static ReadResp.Types.ReadEvent ConvertToReadEvent(ReadReq.Types.Options.Types.UUIDOption uuidOption,
+			ResolvedEvent e) {
+			var readEvent = new ReadResp.Types.ReadEvent {
+				Link = ConvertToRecordedEvent(uuidOption, e.Link, e.LinkPosition?.CommitPosition,
+					e.LinkPosition?.PreparePosition),
+				Event = ConvertToRecordedEvent(uuidOption, e.Event, e.EventPosition?.CommitPosition,
+					e.EventPosition?.PreparePosition),
+			};
+			if (e.OriginalPosition.HasValue) {
+				var position = Position.FromInt64(
+					e.OriginalPosition.Value.CommitPosition,
+					e.OriginalPosition.Value.PreparePosition);
+				readEvent.CommitPosition = position.CommitPosition;
+			} else {
+				readEvent.NoPosition = new Empty();
+			}
+
+			return readEvent;
 		}
 	}
 }
